+++ iOS/iPadDemoApp/pubnub/Classes/View/History/PNChannelHistoryView.m
@@ -261,7 +261,7 @@ typedef enum _PNHistoryMode {
     
     self.traverseSwitch.enabled = !self.pagingModeSwitch.isOn;
     
-    self.messagesCountLimitTextField.text = [NSString stringWithFormat:@"%d", self.historyHelper.maximumNumberOfMessages];
+    self.messagesCountLimitTextField.text = [NSString stringWithFormat:@"%d", (unsigned int)self.historyHelper.maximumNumberOfMessages];
     
     BOOL canUsePagedHistory = (self.historyHelper.channelName && ![self.historyHelper.channelName isEmpty]);
     self.previousPageButton.enabled = canUsePagedHistory;
@@ -498,7 +498,7 @@ typedef enum _PNHistoryMode {
 
 - (NSString *)pickerView:(UIPickerView *)pickerView titleForRow:(NSInteger)row forComponent:(NSInteger)component {
     
-    return [NSString stringWithFormat:@"%d", (row + 1)];
+    return [NSString stringWithFormat:@"%d", (int)(row + 1)];
 }
 
 - (CGFloat)pickerView:(UIPickerView *)pickerView widthForComponent:(NSInteger)component {
